refactor: reuse RadioButton context in RadioButton/Android/IOS/Item (#1223)
