+++ example/src/ExampleList.tsx
@@ -32,7 +32,6 @@ import TextExample from './Examples/TextExample';
 import TextInputExample from './Examples/TextInputExample';
 import ToggleButtonExample from './Examples/ToggleButtonExample';
 import TouchableRippleExample from './Examples/TouchableRippleExample';
-import RadioButtonGroupWithItemsExample from './Examples/RadioButtonGroupWithItemsExample';
 
 type Props = {
   theme: Theme;
@@ -62,7 +61,6 @@ export const examples: { [key: string]: any } = {
   progressbar: ProgressBarExample,
   radio: RadioButtonExample,
   radioGroup: RadioButtonGroupExample,
-  radioGroupWithItems: RadioButtonGroupWithItemsExample,
   searchbar: SearchbarExample,
   snackbar: SnackbarExample,
   surface: SurfaceExample,
