+++ src/components/RadioButton/RadioButton.tsx
@@ -1,22 +1,13 @@
 import * as React from 'react';
 import { Platform } from 'react-native';
-import RadioButtonGroup, {
-  RadioButtonContext,
-  RadioButtonContextType,
-} from './RadioButtonGroup';
-import RadioButtonAndroid, {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  RadioButtonAndroid as _RadioButtonAndroid,
-} from './RadioButtonAndroid';
+import RadioButtonGroup from './RadioButtonGroup';
+import RadioButtonAndroid from './RadioButtonAndroid';
+import RadioButtonIOS from './RadioButtonIOS';
 import RadioButtonItem from './RadioButtonItem';
-import RadioButtonIOS, {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  RadioButtonIOS as _RadioButtonIOS,
-} from './RadioButtonIOS';
-import { withTheme } from '../core/theming';
-import { Theme } from '../types';
+import { withTheme } from '../../core/theming';
+import { Theme } from '../../types';
 
-type Props = {
+export type Props = {
   /**
    * Value of the radio button
    */
@@ -114,33 +105,13 @@ class RadioButton extends React.Component<Props> {
   // @component = ./RadioButtonItem.tsx
   static Item = RadioButtonItem;
 
-  handlePress = (context: RadioButtonContextType) => {
-    const { onPress } = this.props;
-    const onValueChange = context ? context.onValueChange : () => {};
-
-    onPress ? onPress() : onValueChange(this.props.value);
-  };
-
-  isChecked = (context: RadioButtonContextType) =>
-    context.value === this.props.value ? 'checked' : 'unchecked';
-
   render() {
     const Button = Platform.select({
       default: RadioButtonAndroid,
       ios: RadioButtonIOS,
     });
 
-    return (
-      <RadioButtonContext.Consumer>
-        {context => (
-          <Button
-            {...this.props}
-            status={this.props.status || (context && this.isChecked(context))}
-            onPress={() => this.handlePress(context)}
-          />
-        )}
-      </RadioButtonContext.Consumer>
-    );
+    return <Button {...this.props} />;
   }
 }
