+++ src/components/RadioButton/utils.ts
@@ -1,0 +1,27 @@
+export const handlePress = ({
+  onPress,
+  value,
+  onValueChange,
+}: {
+  onPress?: () => void;
+  value: string;
+  onValueChange?: (value: string) => void;
+}) => {
+  onValueChange ? onValueChange(value) : onPress && onPress();
+};
+
+export const isChecked = ({
+  value,
+  status,
+  contextValue,
+}: {
+  value: string;
+  status?: 'checked' | 'unchecked';
+  contextValue?: string;
+}) => {
+  if (contextValue) {
+    return contextValue === value ? 'checked' : 'unchecked';
+  } else {
+    return status;
+  }
+};
